server/requirements.txt
argcomplete==3.5.1
asgiref==3.8.1
blinker==1.8.2
boto3==1.35.54
botocore==1.35.54
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
cfn-flip==1.3.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
dill==0.3.6
dnslib==0.9.25
dnspython==2.7.0
durationpy==0.9
ecdsa==0.19.0
Flask==3.0.3
hjson==3.1.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kappa==0.6.0
localstack==3.8.1
localstack-core==3.8.1
localstack-ext==3.8.1
mangum==0.19.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.1
placebo==0.9.0
plux==1.12.1
psutil==6.1.0
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
Pygments==2.18.0
pyotp==2.9.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
responses==0.25.3
rich==13.9.4
rsa==4.9
s3transfer==0.10.3
semver==3.0.2
six==1.16.0
tabulate==0.9.0
tailer==0.4.1
text-unidecode==1.3
toml==0.10.2
tomli==2.0.2
tqdm==4.66.6
troposphere==4.8.3
typing_extensions==4.12.2
urllib3==1.26.20
watchdog==6.0.0
Werkzeug==3.1.2
xmltodict==0.14.2
zappa==0.59.0
zipp==3.20.2
